+++ CHANGELOG.md
@@ -12,6 +12,17 @@ This project follows [Keep a Changelog](https://keepachangelog.com/) and [Semant
 ### Fixed
 -
 
+## [2.3.7] - 2025-12-03
+### Added
+- Added warning for environment variables logged to console in codebase scanner.
+
+### Changed
+- Updated dependencies to latest versions.
+
+### Fixed
+- Updated jsDocs for better code documentation.
+- Updated some functions for better type safety.
+
 ## [2.3.6] - 2025-12-02
 ### Added
 - Added strict mode handling for framework specific warnings.